signals changed

+++ src/pipes/pipes_utils.c
@@ -6,7 +6,7 @@
 /*   By: ycardona <ycardona@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/24 17:11:45 by mparasku          #+#    #+#             */
-/*   Updated: 2023/08/03 13:35:55 by ycardona         ###   ########.fr       */
+/*   Updated: 2023/08/03 14:41:15 by ycardona         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,15 +56,23 @@ void term_processes(t_data * data)
 	int	j;
 
 	j = 0;
-	while (j <= data->pipe_num)
+	if (data->forked == TRUE)
 	{
-		kill(data->child_pid[j], SIGTERM);
-		j++;
+		while (j <= data->pipe_num)
+		{
+			kill(data->child_pid[j], SIGTERM);
+			j++;
+		}
 	}
 	data->last_exit = 130;
 	g_global->c_kill_child = FALSE;
+	rl_on_new_line();
+	//rl_replace_line("", 0);
+	if (data->forked == FALSE)
+	{
+			rl_redisplay();
+	}
 }
-
 
 void wait_children(t_data *data)
 {
